// src/app/modules/editor/_models/enums/field-type.enum.ts
export enum FieldTypeEnum {
  DOUBLE = 'double',
  STRING = 'string',
  OBJECT = 'object',
  ARRAY = 'array',
  BINARY_DATA = 'binData',
  OBJECT_ID = 'objectId',
  BOOLEAN = 'bool',
  DATE = 'date',
  NULL = 'null',
  REGULAR_EXPRESSION = 'regex',
  JAVASCRIPT = 'javascript',
  JAVASCRIPT_SCOPE = 'javascriptWithScope',
  '32_BIT_INTEGER' = 'int',
  TIMESTAMP = 'timestamp',
  '64_BIT_INTEGER' = 'long',
  DECIMAL_128 = 'decimal',
  MIN_KEY = 'minKey',
  MAX_KEY = 'maxKey'
}